Removed ambiguous Cleanups.java
