+++ src/main/java/org/jabref/logic/importer/util/MetaDataParser.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import org.jabref.logic.cleanup.Cleanups;
+import org.jabref.logic.cleanup.FieldFormatterCleanups;
 import org.jabref.logic.importer.ParseException;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.field.FieldFactory;
@@ -73,7 +73,7 @@ public class MetaDataParser {
                 Path path = Path.of(getSingleItem(value)).normalize();
                 metaData.setLatexFileDirectory(user, path);
             } else if (entry.getKey().equals(MetaData.SAVE_ACTIONS)) {
-                metaData.setSaveActions(Cleanups.parse(value));
+                metaData.setSaveActions(FieldFormatterCleanups.parse(value));
             } else if (entry.getKey().equals(MetaData.DATABASE_TYPE)) {
                 metaData.setMode(BibDatabaseMode.parse(getSingleItem(value)));
             } else if (entry.getKey().equals(MetaData.KEYPATTERNDEFAULT)) {
@@ -111,15 +111,12 @@ public class MetaDataParser {
     /**
      * Returns the first item in the list.
      * If the specified list does not contain exactly one item, then a {@link ParseException} will be thrown.
-     *
-     * @param value
-     * @return
      */
     private static String getSingleItem(List<String> value) throws ParseException {
         if (value.size() == 1) {
             return value.get(0);
         } else {
-            throw new ParseException("Expected a single item but received " + value.toString());
+            throw new ParseException("Expected a single item but received " + value);
         }
     }
