+++ src/test/java/org/jabref/logic/exporter/FieldFormatterCleanupsTest.java
@@ -1,18 +1,24 @@
 package org.jabref.logic.exporter;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-import org.jabref.logic.cleanup.Cleanups;
 import org.jabref.logic.cleanup.FieldFormatterCleanup;
 import org.jabref.logic.cleanup.FieldFormatterCleanups;
 import org.jabref.logic.formatter.IdentityFormatter;
+import org.jabref.logic.formatter.bibtexfields.EscapeAmpersandsFormatter;
+import org.jabref.logic.formatter.bibtexfields.EscapeDollarSignFormatter;
+import org.jabref.logic.formatter.bibtexfields.EscapeUnderscoresFormatter;
+import org.jabref.logic.formatter.bibtexfields.LatexCleanupFormatter;
 import org.jabref.logic.formatter.bibtexfields.NormalizeDateFormatter;
+import org.jabref.logic.formatter.bibtexfields.NormalizeMonthFormatter;
 import org.jabref.logic.formatter.bibtexfields.NormalizePagesFormatter;
 import org.jabref.logic.formatter.casechanger.LowerCaseFormatter;
 import org.jabref.model.entry.BibEntry;
+import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.types.StandardEntryType;
 
@@ -41,7 +47,7 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void checkSimpleUseCase() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("title[identity]"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("title[identity]"));
 
         FieldFormatterCleanup identityInTitle = new FieldFormatterCleanup(StandardField.TITLE, new IdentityFormatter());
         assertEquals(Collections.singletonList(identityInTitle), actions.getConfiguredActions());
@@ -53,7 +59,7 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void invalidSaveActionSting() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("title"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("title"));
 
         assertEquals(Collections.emptyList(), actions.getConfiguredActions());
 
@@ -64,7 +70,7 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void checkLowerCaseSaveAction() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("title[lower_case]"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("title[lower_case]"));
 
         FieldFormatterCleanup lowerCaseTitle = new FieldFormatterCleanup(StandardField.TITLE, new LowerCaseFormatter());
         assertEquals(Collections.singletonList(lowerCaseTitle), actions.getConfiguredActions());
@@ -76,7 +82,7 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void checkTwoSaveActionsForOneField() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("title[lower_case,identity]"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("title[lower_case,identity]"));
 
         FieldFormatterCleanup lowerCaseTitle = new FieldFormatterCleanup(StandardField.TITLE, new LowerCaseFormatter());
         FieldFormatterCleanup identityInTitle = new FieldFormatterCleanup(StandardField.TITLE, new IdentityFormatter());
@@ -89,7 +95,7 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void checkThreeSaveActionsForOneField() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("title[lower_case,identity,normalize_date]"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("title[lower_case,identity,normalize_date]"));
 
         FieldFormatterCleanup lowerCaseTitle = new FieldFormatterCleanup(StandardField.TITLE, new LowerCaseFormatter());
         FieldFormatterCleanup identityInTitle = new FieldFormatterCleanup(StandardField.TITLE, new IdentityFormatter());
@@ -103,7 +109,7 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void checkMultipleSaveActions() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("pages[normalize_page_numbers]title[lower_case]"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("pages[normalize_page_numbers]title[lower_case]"));
         List<FieldFormatterCleanup> formatterCleanups = actions.getConfiguredActions();
 
         FieldFormatterCleanup normalizePages = new FieldFormatterCleanup(StandardField.PAGES, new NormalizePagesFormatter());
@@ -119,7 +125,7 @@ public class FieldFormatterCleanupsTest {
     @Test
     public void checkMultipleSaveActionsWithMultipleFormatters() {
         FieldFormatterCleanups actions = new FieldFormatterCleanups(true,
-                Cleanups.parse("pages[normalize_page_numbers,normalize_date]title[lower_case]"));
+                FieldFormatterCleanups.parse("pages[normalize_page_numbers,normalize_date]title[lower_case]"));
         List<FieldFormatterCleanup> formatterCleanups = actions.getConfiguredActions();
 
         FieldFormatterCleanup normalizePages = new FieldFormatterCleanup(StandardField.PAGES, new NormalizePagesFormatter());
@@ -135,9 +141,173 @@ public class FieldFormatterCleanupsTest {
 
     @Test
     public void clearFormatterRemovesField() {
-        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, Cleanups.parse("month[clear]"));
+        FieldFormatterCleanups actions = new FieldFormatterCleanups(true, FieldFormatterCleanups.parse("month[clear]"));
         actions.applySaveActions(entry);
 
         assertEquals(Optional.empty(), entry.getField(StandardField.MONTH));
+    }
+
+
+    @Test
+    void parserKeepsSaveActions() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                month[normalize_month]
+                pages[normalize_page_numbers]
+                title[escapeAmpersands,escapeDollarSign,escapeUnderscores,latex_cleanup]
+                booktitle[escapeAmpersands,escapeDollarSign,escapeUnderscores,latex_cleanup]
+                publisher[escapeAmpersands,escapeDollarSign,escapeUnderscores,latex_cleanup]
+                journal[escapeAmpersands,escapeDollarSign,escapeUnderscores,latex_cleanup]
+                abstract[escapeAmpersands,escapeDollarSign,escapeUnderscores,latex_cleanup]
+                """);
+
+        List<FieldFormatterCleanup> expected = new ArrayList<>(30);
+        expected.add(new FieldFormatterCleanup(StandardField.MONTH, new NormalizeMonthFormatter()));
+        expected.add(new FieldFormatterCleanup(StandardField.PAGES, new NormalizePagesFormatter()));
+        for (Field field : List.of(StandardField.TITLE, StandardField.BOOKTITLE, StandardField.PUBLISHER, StandardField.JOURNAL, StandardField.ABSTRACT)) {
+            expected.add(new FieldFormatterCleanup(field, new EscapeAmpersandsFormatter()));
+            expected.add(new FieldFormatterCleanup(field, new EscapeDollarSignFormatter()));
+            expected.add(new FieldFormatterCleanup(field, new EscapeUnderscoresFormatter()));
+            expected.add(new FieldFormatterCleanup(field, new LatexCleanupFormatter()));
+        }
+
+        assertEquals(expected, fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserParsesLatexCleanupFormatter() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[latex_cleanup]
+                """);
+        assertEquals(
+                List.of(new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter())),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserParsesTwoFormatters() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[escapeUnderscores,latex_cleanup]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter())
+                ),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserParsesFourFormatters() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[escapeAmpersands,escapeDollarSign,escapeUnderscores,latex_cleanup]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeDollarSignFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter())
+                ),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserParsesTwoFormattersWithCommas() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[escapeUnderscores,latex_cleanup]
+                booktitle[escapeAmpersands,escapeDollarSign]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeDollarSignFormatter())
+                ),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserParsesTwoFormattersOneWithComma() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                pages[normalize_page_numbers]
+                booktitle[escapeAmpersands,escapeDollarSign]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.PAGES, new NormalizePagesFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeDollarSignFormatter())
+                ),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserParsesThreeFormattersTwoWithComma() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                pages[normalize_page_numbers]
+                title[escapeUnderscores,latex_cleanup]
+                booktitle[escapeAmpersands,escapeDollarSign]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.PAGES, new NormalizePagesFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeDollarSignFormatter())
+                ),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserWithTwoAndThree() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[escapeAmpersands,escapeUnderscores,latex_cleanup]
+                booktitle[escapeAmpersands,escapeUnderscores,latex_cleanup]
+                """);
+
+        List<FieldFormatterCleanup> expected = new ArrayList<>(30);
+        for (Field field : List.of(StandardField.TITLE, StandardField.BOOKTITLE)) {
+            expected.add(new FieldFormatterCleanup(field, new EscapeAmpersandsFormatter()));
+            expected.add(new FieldFormatterCleanup(field, new EscapeUnderscoresFormatter()));
+            expected.add(new FieldFormatterCleanup(field, new LatexCleanupFormatter()));
+        }
+
+        assertEquals(expected, fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserWithFourEntries() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[escapeUnderscores,latex_cleanup]
+                booktitle[escapeAmpersands,escapeUnderscores,latex_cleanup]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new LatexCleanupFormatter())
+                ),
+                fieldFormatterCleanups);
+    }
+
+    @Test
+    void parserTest() {
+        List<FieldFormatterCleanup> fieldFormatterCleanups = FieldFormatterCleanups.parse("""
+                title[escapeAmpersands,escapeUnderscores,latex_cleanup]
+                booktitle[escapeAmpersands,latex_cleanup]
+                """);
+        assertEquals(
+                List.of(
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new EscapeUnderscoresFormatter()),
+                        new FieldFormatterCleanup(StandardField.TITLE, new LatexCleanupFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new EscapeAmpersandsFormatter()),
+                        new FieldFormatterCleanup(StandardField.BOOKTITLE, new LatexCleanupFormatter())
+                ),
+                fieldFormatterCleanups);
     }
 }
